Move private user out of getDestinationsForUser
